refactor(graph-gateway): move legacy_auth_adapter under client_query module
