+++ graph-gateway/Cargo.toml
@@ -46,6 +46,7 @@ tracing-subscriber.workspace = true
 hickory-resolver = "0.24.0"
 uuid = { version = "1.6", default-features = false, features = ["v4"] }
 axum.workspace = true
+headers = "0.3.9"
 
 [dev-dependencies]
 assert_matches = "1.5.0"
